working - Monitor in separate class
