+++ yadi/dss/load_shape.py
@@ -1,0 +1,1 @@
+load_shape.py
