+++ yadi/dss/qsts.py
@@ -9,7 +9,7 @@ October 6th, 2021
 from tkinter import W
 import numpy as np
 import pandas as pd
-import yadi.yadi.dss.model as model 
+import yadi.yadi.dss.monitor as monitor 
 import warnings
 from tqdm import tqdm
 import pathlib
@@ -19,7 +19,7 @@ from calendar import monthrange
 #Optional: Turn off complex value warnings
 #warnings.simplefilter("ignore", np.ComplexWarning)
 
-class DSS_Timeseries(model.DSS_Data):
+class DSS_Timeseries(monitor.DSS_Monitor):
 
     def __init__(
             self,
@@ -50,7 +50,7 @@ class DSS_Timeseries(model.DSS_Data):
             - verbose (boolean): whether or not to print verbose logs
             - per_unit (boolean): whether or not to return per unit values at all times (default is False) #TODO
         """
-        super().__init__(redirects,precompile=precompile)
+        super().__init__(redirects, precompile=precompile)
 
         #--- Simulation parameters ---#
         self.time_step = time_step
@@ -281,30 +281,6 @@ class DSS_Timeseries(model.DSS_Data):
         return D_diff_N
 
 
-    def __export_monitor(self,monitor_name,verbose=False):
-        """Exports a single monitor to a dataframe"""
-        err = self.dss.run_command("export monitors {monitor_name}".format(
-                monitor_name=monitor_name
-            )
-        )
-        print('Monitor {monitor_name} Export Returned {err}'.format(
-            monitor_name=monitor_name,
-            err=err
-        ))
-        monitor_info = self.dss.utils.monitors_to_dataframe() #Get the monitor info df
-        #Get the monitor info df index for monitor_name
-        monitor_index = monitor_info.index.get_loc(monitor_info.index[monitor_info.index == monitor_name][0])
-        
-        #Make timeseries_df from exported csv
-        timeseries_df = pd.read_csv(monitor_info['FileName'][monitor_index],sep=r'\s*,\s*',
-                        header=0, encoding='ascii', engine='python')
-        if(verbose):
-            print('monitor_name is: ',monitor_name)
-            print('monitor_idex is: ',monitor_index)
-        return timeseries_df
-
-
-
     def set_loadshape(self,loadshape_path,loadshape_name='loadshape1'):
         """
         Sets a loadshape for all loads
@@ -322,17 +298,18 @@ class DSS_Timeseries(model.DSS_Data):
         ) #change all loads            
     
 
-    def __check_qsts_initialization(self):
+    def __check_qsts_initialization(self, native=False):
         """Check if QSTS has been properly initialized"""
          #Check to see if QSTS is initialized
         if(not self.__qsts_initialized):
             warnings.warn("QSTS has not been initialized. Initiailizing before run.")
             self.compile_dss()
-            self.initialize_qsts()
+            self.initialize_qsts(native)
+
         elif(self.__qsts_complete):
             # warnings.warn("QSTS has already been run for the input files. Recompiling before run...")
             self.compile_dss()
-            self.initialize_qsts()
+            self.initialize_qsts(native)
 
     #  #################################################
     #  ######### native Opendss QSTS #########
@@ -402,7 +379,7 @@ class DSS_Timeseries(model.DSS_Data):
                 thermalLimitDict[line] = self.dss.Lines.NormAmps()
         return pd.Series(thermalLimitDict)
 
-    def initialize_qsts(self, monitor_trafos=True, monitor_lines=True, monitor_loads=True, verbose=False):
+    def initialize_qsts(self, native, verbose=False):
         """
         Initialize a chosen-mode Quasi-Static Time Series simulation.
 
@@ -412,30 +389,30 @@ class DSS_Timeseries(model.DSS_Data):
         """
         errs = []
 
-        if(monitor_loads):
-            self.__set_monitor_all_loads(verbose=verbose)
+        if(native):
+            number = self.simulation_steps
 
-        if(monitor_lines):
-            self.__set_monitor_all_lines(verbose=verbose)
+            self.set_monitor_all_loads(verbose=verbose)
 
-        if(monitor_trafos):
-            self.__set_monitor_all_trafos(verbose=verbose)
+            self.set_monitor_all_lines(verbose=verbose)
+
+            self.set_monitor_all_trafos(verbose=verbose)
+        else:
+            number = self.solution_number
 
         errs.append(
             self.dss.run_command(f'Set controlmode={self.simulation_controlmode}')
         )
         errs.append(
             self.dss.run_command(f"Set mode={self.simulation_mode} "
-                                 f"number={self.solution_number} "
+                                 f"number={number} "
                                  f"stepsize={self.time_step} "
                                  f"maxcontroliter={self.maxcontroliter} "
                                  f"maxiterations={self.maxiterations} "
                                  f"miniterations={self.miniterations} "
                                  )
         )
-        errs.append(
-            self.dss.run_command('Set maxcontroliter=600')
-        )
+
         # print('QSTS Initialized, Returned: ', [err for err in errs])
         self.__qsts_initialized = True
 
@@ -450,9 +427,7 @@ class DSS_Timeseries(model.DSS_Data):
         """
         if userDemand is not None:
             self.__setAllLoadShapes(userDemand[0], userDemand[1])
-            self.offset = 0
-        else:
-            self.offset = 3  # the dss duty lenght is shifted!
+
         # run routine with modified loadShapes
         self.__run_qsts_OpenDSS_duty()
 
@@ -498,13 +473,10 @@ class DSS_Timeseries(model.DSS_Data):
                 self.__modifyLoadShapePQ(Pmult, Qmult, loadShapeName)
             else:  # not a load
                 self.dss.LoadShape.Name(loadShapeName)
-                offset = 3 
                 Pmult = self.dss.LoadShape.PMult()
-                Pmult = np.asarray(Pmult[offset:])
                 self.__modifyLoadShapeP(tuple(Pmult), loadShapeName)
             # check load modification
             # self.dss.LoadShape.Name(loadShapeName)
-            # offset = 0
             # Pmult = self.dss.LoadShape.PMult()
             # Qmult = self.dss.LoadShape.QMult()
 
@@ -592,233 +564,34 @@ class DSS_Timeseries(model.DSS_Data):
         self.kvarDemand = kvarDemand
         return kwDemand, kvarDemand
 
-    def __set_monitor(self, element_name, element_type, mon_name_prefix, power=True, voltage=True, verbose=False):
-        """Sets a monitor on element_name of element_type"""
-        if(power):  # If power monitor is enabled
-            mon_name = mon_name_prefix + "power"
-            err1 = self.dss.run_command(f"New Monitor.{mon_name} "
-                                        f"Element={element_type}.{element_name} "
-                                        "terminal=1 PPolar=no mode=1")
-            if(verbose):
-                print("Monitor type: ", mon_name_prefix + "power", " placed on", element_type, " name: ", element_name, "with errors: ", err1)
-
-        if(voltage):
-            mon_name = mon_name_prefix + "voltage"
-            err2 = self.dss.run_command(f"New Monitor.{mon_name} "
-                                        f"Element={element_type}.{element_name} "
-                                        "terminal=1 vipolar=yes mode=0")
-            if(verbose):
-                print("Monitor type: ", mon_name_prefix + "voltage", " placed on", element_type, " name: ", element_name, "with errors: ", err2)
 
     def __run_qsts_OpenDSS_duty(self):
         """
         Run a "Quasi-Static Time-Series" and get multivariate timeseries dataets of voltage magnitudes, complex powers, and currents, for each node
         """
         # Check QSTS initialization
-        self.__check_qsts_initialization()
+        self.__check_qsts_initialization(native=True)
+
         # Run Duty mode qsts
-        # whole duty
         self.dss.Text.Command('solve')
-        voltage_profiles, kw_profiles, kvar_profiles = self.__get_monitor_all_loads()
-        lineIjk, linePjk, lineQjk = self.__get_monitor_all_lines()
-        trafoIjk, trafoPjk, trafoQjk = self.__get_monitor_all_trafos()
+
+        # whole duty
+        voltage_profiles, kw_profiles, kvar_profiles = self.get_monitor_all_loads()
+        lineIjk, linePjk, lineQjk = self.get_monitor_all_lines()
+        trafoIjk, trafoPjk, trafoQjk = self.get_monitor_all_trafos()
         self.__qsts_complete = True
+
         # load quantities
         self.loadVolts = voltage_profiles
         self.loadKws = kw_profiles
         self.loadKvars = kvar_profiles
+
         # line quantities
         self.lineIjks = lineIjk
         self.linePjks = linePjk
         self.lineQjk = lineQjk
+
         # trafo quantities
         self.trafoIjks = trafoIjk
         self.trafoPjks = trafoPjk
         self.trafoQjk = trafoQjk
-
-    def __set_monitor_all_lines(self, verbose=False):
-        """Sets timeseries power monitors on all lines before solving"""
-        lines = self.dss.Lines.AllNames()
-        for n, line_name in enumerate(lines):
-            mon_name_prefix = "mon_" + str(line_name) + "_"
-            self.__set_monitor(element_name=line_name, element_type="Line",
-                               mon_name_prefix=mon_name_prefix, power=True,
-                               voltage=True, verbose=verbose)
-
-    def __set_monitor_all_trafos(self, verbose=False):
-        """Sets timeseries power monitors on all trafos before solving"""
-        trafos = self.dss.Transformers.AllNames()
-        for n, trafo_name in enumerate(trafos):
-            mon_name_prefix = "mon_" + str(trafo_name) + "_"
-            self.__set_monitor(element_name=trafo_name, element_type="Transformer",
-                               mon_name_prefix=mon_name_prefix, power=True,
-                               voltage=True, verbose=verbose)
-
-    def __set_monitor_all_loads(self, verbose=False):
-        """Sets timeseries power monitors on all loads before solving"""
-        loads = self.dss.Loads.AllNames()
-        for n, load_name in enumerate(loads):
-            mon_name_prefix = "mon_" + str(load_name) + "_"
-            self.__set_monitor(element_name=load_name, element_type="Load",
-                               mon_name_prefix=mon_name_prefix, power=True,
-                               voltage=True, verbose=verbose)
-
-    def __get_monitor_all_trafos(self, verbose=False):
-        """Sets timeseries power monitors on all loads before solving"""
-        trafos = self.dss.Transformers.AllNames()
-        Ijk_dict = dict()
-        Pjk_dict = dict()
-        Qjk_dict = dict()
-        for n, trafo_name in enumerate(trafos):
-            Ijk, Pjk, Qjk = self.__get_monitor_timeseries(name=trafo_name, type="Transformer")
-            if (len(Ijk.shape) > 1) and (len(Pjk.shape) > 1) and (len(Qjk.shape) > 1):
-                self.dss.Transformers.Name(trafo_name)  # set current trafo as active
-                phases = self.dss.CktElement.NumPhases()
-                for ph in range(phases):
-                    Ijk_dict[trafo_name + f".{ph + 1}"] = Ijk[:, ph]
-                    Pjk_dict[trafo_name + f".{ph + 1}"] = Pjk[:, ph]
-                    Qjk_dict[trafo_name + f".{ph + 1}"] = Qjk[:, ph]
-            else:
-                Ijk_dict[trafo_name] = Ijk
-                Pjk_dict[trafo_name] = Pjk[:, 0]
-                Qjk_dict[trafo_name] = Qjk[:, 0]
-        
-        Ijk_profiles = pd.DataFrame.from_dict(Ijk_dict)
-        Ijk_profiles = Ijk_profiles.iloc[self.offset:, :]
-        Pjk_profiles = pd.DataFrame.from_dict(Pjk_dict)
-        Pjk_profiles = Pjk_profiles.iloc[self.offset:, :]
-        Qjk_profiles = pd.DataFrame.from_dict(Qjk_dict)
-        Qjk_profiles = Qjk_profiles.iloc[self.offset:, :]
-
-        dt_index = pd.date_range(start='1/1/2019', periods=self.simulation_steps - self.offset, freq='H')
-        Ijk_profiles = Ijk_profiles.set_index(dt_index)
-        Pjk_profiles = Pjk_profiles.set_index(dt_index)
-        Qjk_profiles = Qjk_profiles.set_index(dt_index)
-        return Ijk_profiles, Pjk_profiles, Qjk_profiles
-
-    def __get_monitor_all_lines(self, verbose=False):
-        """Sets timeseries power monitors on all loads before solving"""
-        lines = self.dss.Lines.AllNames()
-        Ijk_dict = dict()
-        Pjk_dict = dict()
-        Qjk_dict = dict()
-        for n, line_name in enumerate(lines):
-            Ijk, Pjk, Qjk = self.__get_monitor_timeseries(name=line_name, type="Line")
-            if (len(Ijk.shape) > 1) and (len(Pjk.shape) > 1) and (len(Qjk.shape) > 1):
-                self.dss.Lines.Name(line_name)  # set current line as active
-                phases = self.dss.Lines.Phases()
-                for ph in range(phases):
-                    Ijk_dict[line_name + f".{ph + 1}"] = Ijk[:, ph]
-                    Pjk_dict[line_name + f".{ph + 1}"] = Pjk[:, ph]
-                    Qjk_dict[line_name + f".{ph + 1}"] = Qjk[:, ph]
-            else:
-                Ijk_dict[line_name] = Ijk
-                Pjk_dict[line_name] = Pjk[:, 0]
-                Qjk_dict[line_name] = Qjk[:, 0]
-        
-        Ijk_profiles = pd.DataFrame.from_dict(Ijk_dict)
-        Ijk_profiles = Ijk_profiles.iloc[self.offset:, :]
-        Pjk_profiles = pd.DataFrame.from_dict(Pjk_dict)
-        Pjk_profiles = Pjk_profiles.iloc[self.offset:, :]
-        Qjk_profiles = pd.DataFrame.from_dict(Qjk_dict)
-        Qjk_profiles = Qjk_profiles.iloc[self.offset:, :]
-
-        dt_index = pd.date_range(start='1/1/2019', periods=self.simulation_steps - self.offset, freq='H')
-        Ijk_profiles = Ijk_profiles.set_index(dt_index)
-        Pjk_profiles = Pjk_profiles.set_index(dt_index)
-        Qjk_profiles = Qjk_profiles.set_index(dt_index)
-        return Ijk_profiles, Pjk_profiles, Qjk_profiles
-
-    def __get_monitor_all_loads(self, verbose=False):
-        """Sets timeseries power monitors on all loads before solving"""
-        loads = self.dss.Loads.AllNames()
-        voltage_dict = dict()
-        kw_dict = dict()
-        kvar_dict = dict()
-        for n, load_name in enumerate(loads):
-            volts, kws, kvars = self.__get_monitor_timeseries(load_name)
-            voltage_dict[load_name] = volts
-            kw_dict[load_name] = kws
-            kvar_dict[load_name] = kvars
-        
-        voltage_profiles = pd.DataFrame.from_dict(voltage_dict)
-        voltage_profiles = voltage_profiles.iloc[self.offset:, :]
-        kw_profiles = pd.DataFrame.from_dict(kw_dict)
-        kw_profiles = kw_profiles.iloc[self.offset:, :]
-        kvar_profiles = pd.DataFrame.from_dict(kvar_dict)
-        kvar_profiles = kvar_profiles.iloc[self.offset:, :]
-
-        dt_index = pd.date_range(start='1/1/2019', periods=self.simulation_steps - self.offset, freq='H')
-        voltage_profiles = voltage_profiles.set_index(dt_index)
-        kw_profiles = kw_profiles.set_index(dt_index)
-        kvar_profiles = kvar_profiles.set_index(dt_index)
-        return voltage_profiles, kw_profiles, kvar_profiles
-
-    def __get_monitor_timeseries(self, name, type="Load"):
-        """
-        Gets the voltage, active, and reactive power timeseries dictionary for a single elemented in the system:
-        {(V_i,t,P_i,t,Q_i)_t}_{t=1,..,M}
-
-        Parameters:
-        ---
-            dss: the dss object
-            element: name of the lement
-        """
-        if type == "Line" or type == "Transformer":
-            Ijk_ts = self.__export_monitor_voltage(name, type)
-            Pjk_ts, Qjk_ts = self.__export_monitor_power(name, type)
-            return Ijk_ts, Pjk_ts, Qjk_ts
-        else:
-            voltage_ts = self.__export_monitor_voltage(name, type)
-            kw_ts, kvar_ts = self.__export_monitor_power(name, type)
-            return voltage_ts, kw_ts, kvar_ts
-
-    def __export_monitor_voltage(self, name, type):
-        """Gets the voltage timeseries for an element that is monitored"""
-        monitor_name = "mon_" + name + "_voltage"
-        # set the active monitor according to name
-        self.dss.Monitors.Name(monitor_name)
-        voltage_matrix = self.dss.Monitors.AsMatrix()  # N timesteps x M chanels (t, 0, v1, angle 1, ... I1, angle1, ...)
-        if type == "Load":
-            return voltage_matrix[:, 2]  # interested in v1 for loads
-        elif type == "Line":
-            self.dss.Lines.Name(name)  # set current line as active
-            phases = self.dss.Lines.Phases()
-            if phases == 1:
-                # numCols = voltage_matrix.shape[1]
-                # print(f"phase:{phases}-cols:{numCols}")
-                return voltage_matrix[:, 4]  # interested in current magnitudes for the lines
-            if phases == 3:
-                # numCols = voltage_matrix.shape[1]
-                # print(f"phase:{phases}-cols:{numCols}")
-                return voltage_matrix[:, 8::2]  # interested in current magnitudes for the lines
-        elif type == "Transformer":
-            self.dss.Transformers.Name(name)  # set current line as active
-            # self.dss.Circuit.SetActiveElement(type + "." + name)
-            phases = self.dss.CktElement.NumPhases()
-            if phases == 1:
-                # numCols = voltage_matrix.shape[1]
-                # print(f"phase:{phases}-cols:{numCols}")
-                return voltage_matrix[:, 6]  # interested in current magnitudes for the lines
-            if phases == 3:
-                # numCols = voltage_matrix.shape[1]
-                # print(f"phase:{phases}-cols:{numCols}")
-                return voltage_matrix[:, 10::2]  # interested in current magnitudes for the lines
-        
-
-    def __export_monitor_power(self, name, type):
-        """Gets the active and reactive power timeseries for an element monitored"""
-        monitor_name = "mon_" + name + "_power"
-        # set the active monitor according to name
-        self.dss.Monitors.Name(monitor_name)
-        power_matrix = self.dss.Monitors.AsMatrix()  # N timesteps x M chanels (t, 0, P1, Q1, ....)
-        if type == "Load":
-            return power_matrix[:, 2], power_matrix[:, 3]  # interesteed in P1, Q1 for loads
-        elif type == "Line":
-            # numCols = power_matrix.shape[1]
-            # print(f"cols:{numCols}")
-            return power_matrix[:, 2::2], power_matrix[:, 3::2]   # interesteed in Pjks and Qjks for the phases
-        elif type == "Transformer":
-            # numCols = power_matrix.shape[1]
-            # print(f"cols:{numCols}")
-            return power_matrix[:, 2::2], power_matrix[:, 3::2]   # interesteed in Pjks and Qjks for the phases
